// host_and_shared/late-mate-cli/frontend/src/generated/types/HidReport.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { KeyboardReport } from "./KeyboardReport";
import type { MouseReport } from "./MouseReport";

export type HidReport =
  | ({ type: "mouse" } & MouseReport)
  | ({ type: "keyboard" } & KeyboardReport);